[fxfs] Change reader counts to u64
Change `current_count` in `Inner` and `count` in `EpochEraseList`
from `u16` to `u64`. This prevents an integer overflow when there
are many concurrent readers, which could lead to a use-after-free
vulnerability.

The `u16` type was likely a legacy holdover.

Fixed: 517452957
Test: fxfs-test
TAG: agy
CONV: 920ec6fa-3a70-4cfe-a1ad-fcb91c96a37b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1684654
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>

+++ src/storage/fxfs/src/lsm_tree/skip_list_layer.rs
@@ -85,7 +85,7 @@ struct Inner<K, V> {
     epoch: u64,
 
     // The number of readers on the current epoch.
-    current_count: u16,
+    current_count: u64,
 
     // A list of nodes to be freed once the read counts have reached zero.
     erase_lists: BTreeMap<u64, EpochEraseList<K, V>>,
@@ -101,7 +101,7 @@ struct Inner<K, V> {
 struct EpochEraseList<K, V> {
     // The number of readers still associated with this epoch.  When this reaches zero, the list can
     // be freed once all previous epochs have been freed.
-    count: u16,
+    count: u64,
     // We represent the list by storing the head and tail of the list which each node chained to the
     // next.
     start: NonNull<SkipListNode<K, V>>,
